move UserService from service->webservice package
