+++ src/test/scala/org/broadinstitute/dsde/firecloud/service/UserServiceSpec.scala
@@ -5,7 +5,7 @@ import org.broadinstitute.dsde.firecloud.mock.MockUtils
 import org.broadinstitute.dsde.firecloud.mock.MockUtils._
 import org.broadinstitute.dsde.firecloud.model.ModelJsonProtocol._
 import org.broadinstitute.dsde.firecloud.model._
-import org.broadinstitute.dsde.firecloud.webservice.RegisterApiService
+import org.broadinstitute.dsde.firecloud.webservice.{RegisterApiService, UserService}
 import org.mockserver.integration.ClientAndServer
 import org.mockserver.integration.ClientAndServer._
 import org.mockserver.model.HttpRequest._
